An encryptor,decryptor pair(user_inp_num_resp)
An encrypter, decryptor pair. The user inputs the path of the storage folder for the serialized files. The user also inputs the number of questions and clients that are to be simulated. The decryptor needs input to locate the path of the storage folder for the serialized files. The user also needs to input the number of responses it wants averaged. Depending on the input, the program will average the first 'input' responses. The pair of programs is designed to work on a single system.